Add windows support for go-init
