+++ init/cli/lib.go
@@ -21,9 +21,7 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"syscall"
 
-	ps "github.com/mitchellh/go-ps"
 	"github.com/palantir/go-java-launcher/launchlib"
 	"github.com/palantir/pkg/cli"
 	"github.com/pkg/errors"
@@ -150,8 +148,10 @@ func getConfiguredCommands(ctx cli.Context, loggers launchlib.ServiceLoggers) (m
 }
 
 func isPidRunning(pid int) (bool, *os.Process, error) {
-	// Docs say FindProcess always succeeds on Unix.
-	proc, _ := os.FindProcess(pid)
+	proc, err := os.FindProcess(pid)
+	if err != nil {
+		return false, nil, err
+	}
 	running, err := isProcRunning(proc)
 	if err != nil {
 		return false, nil, err
@@ -160,26 +160,4 @@ func isPidRunning(pid int) (bool, *os.Process, error) {
 		return true, proc, nil
 	}
 	return false, nil, nil
-}
-
-func isProcRunning(proc *os.Process) (bool, error) {
-	// This is the way to check if a process exists: https://linux.die.net/man/2/kill.
-	// On linux, this may respond true if there is a thread running with the same id.
-	running := proc.Signal(syscall.Signal(0)) == nil
-	if !running {
-		return false, nil
-	}
-
-	// On linux, iterating over processes will only return running processes. Unfortunately,
-	// getting exactly the one pid will also return a thread of the same id.
-	procs, err := ps.Processes()
-	if err != nil {
-		return false, err
-	}
-	for _, p := range procs {
-		if p.Pid() == proc.Pid {
-			return true, nil
-		}
-	}
-	return false, nil
 }
